rename layout to bloglayout
