+++ src/components/blogLayout.js
